+++ app/src/main/res/layout/fragment_edit_spec_route.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Frags.FragEditSpecRoute" >
+    tools:context=".Editables.FragEditSpecRoute" >
 
     <ImageView
         android:id="@+id/spRClimb"
