+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--activity_main_drawer.xml-->
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     tools:showIn="navigation_view">
@@ -16,6 +17,9 @@
             android:id="@+id/nav_slideshow"
             android:icon="@drawable/ic_menu_slideshow"
             android:title="@string/menu_slideshow" />
-
+        <item
+            android:id="@+id/nav_admin"
+            android:icon="@drawable/ic_launcher_background"
+            android:title="Admin"/>
     </group>
 </menu>
